refactor(util): move use-cached-resource from teams/ to util/
