+++ shared/teams/use-teams-list.tsx
@@ -14,7 +14,7 @@ import {
   type CachedResourceCache,
   createCachedResourceCache,
   useCachedResource,
-} from './use-cached-resource'
+} from '@/util/use-cached-resource'
 
 type TeamsList = {
   reload: () => void
